refactor(tooltip): migrate to TypeScript
